requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.3
distlib==0.3.3
dnspython==2.1.0
filelock==3.3.1
Flask==2.0.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
platformdirs==2.4.0
pygame==2.1.0
pymongo==4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
six==1.15.0
style==1.1.0
update==0.0.1
urllib3==1.26.6
virtualenv==20.8.1
Werkzeug==2.0.2
